poly-commitment: move batch IPA test into ipa_commitment
I guess it has been initially called batch_15_wires to test the Kimchi setting
that uses 15 columns. However, the code does not test 15 wires, but simply a
batch of polynomials, with blinders. It is different than the test above in
ipa_commitment.rs that only tests two polynomials.
